% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_mutation.R
\name{add_mutation}
\alias{add_mutation}
\title{Title}
\usage{
add_mutation(Eset, mutations)
}
\arguments{
\item{Eset}{Expression Set where the mutations will be added as a column in the pData table}

\item{mutations}{a vector of mutations to be added}
}
\value{
Expression Set where the mutations will are added as a column in the pData table
}
\description{
Title
}
\examples{
\dontrun{
add_mutation(Eset = Eset,
mutations = c("VHL", "PBRM1")
)
}
}
