% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build_TCGA_Eset.R
\name{build_TCGA_Eset}
\alias{build_TCGA_Eset}
\title{Create an Expression Set from TCGA Data, downloaded from https://gdac.broadinstitute.org/}
\usage{
build_TCGA_Eset(clinical_file, expression_file, source, skip = FALSE)
}
\arguments{
\item{clinical_file}{PATH of the clinical data from TCGA, e.g.: "../../../Projects/TCGA-Elisa/Clinical/LAML.clin.merged.txt"}

\item{expression_file}{PATH of the expression data from TCGA, e.g.: "../../../Projects/TCGA-Elisa/RNA/LAML.rnaseqv2__illuminahiseq_rnaseqv2__unc_edu__Level_3__RSEM_genes_normalized__data.data.txt"}

\item{source}{Define the source of the data. "cBio" and "cBio_z_score" for cBioportal Data, "firehose" for Firehose data.}

\item{skip}{Please Check in the clinical file in which row the header column appears. The amount of rows skiped should be entered as e.g. skip = 5.}
}
\value{
ExpressionSet
}
\description{
Create an Expression Set from TCGA Data, downloaded from https://gdac.broadinstitute.org/
}
\examples{
\dontrun{
build_TCGA_Eset(clinical_file <- "../../../Projects/TCGA-Elisa/tcga_KIRC_cBio/data_bcr_clinical_data_patient.txt",
expression_file <- "../../../Projects/TCGA-Elisa/tcga_KIRC_cBio/Kdata_RNA_Seq_v2_mRNA_median_Zscores.txt",
source = "cBio_z_score",
normalization = FALSE)
}

}
