% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fetch_TCGA_mutations.R
\name{fetch_TCGA_mutations}
\alias{fetch_TCGA_mutations}
\title{Add an Mutation Column to an Expression set made by build_TCGA_Eset() from TCGA Mutation Data, downloaded from https://gdac.broadinstitute.org/
Please download the Mutation_Packager_Oncotated_Calls}
\usage{
fetch_TCGA_mutations(path, Eset)
}
\arguments{
\item{path}{Path to the mutation Files downloaded from Firehose TCGA}

\item{Eset}{Eset that will be used}
}
\value{
Returns the Expression Set with added mutation data
}
\description{
Add an Mutation Column to an Expression set made by build_TCGA_Eset() from TCGA Mutation Data, downloaded from https://gdac.broadinstitute.org/
Please download the Mutation_Packager_Oncotated_Calls
}
\examples{
mutset <- fetch_TCGA_mutations(path = "../../../Projects/TCGA-Elisa/KIRC.Mutation_Packager_Oncotated_Calls/", Eset = Eset)



}
