% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hazard_list.R
\name{hazard_list}
\alias{hazard_list}
\title{Title}
\usage{
hazard_list(x, Eset, additional, exclude)
}
\arguments{
\item{x}{a character vector with 1 or more gene names}

\item{Eset}{An Expression Set}

\item{additional}{covariate to subset groups}

\item{exclude}{Factors of the covariates to exclude}
}
\value{
A List of Hazard Ratios of a list of genes
}
\description{
Title
}
\examples{
\dontrun{
hazard_list(x = c("FOXA2"), Eset = Eset, additional = "pathologic_stage", exclude = c("Stage IV")
}
}
