% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/list_clinical.R
\name{list_clinical}
\alias{list_clinical}
\title{Small function that lists all the available clinical features of a given Expressionset}
\usage{
list_clinical(Eset)
}
\arguments{
\item{Eset}{ExpressionSet}
}
\value{
list of all available clinical features
}
\description{
Small function that lists all the available clinical features of a given Expressionset
}
\examples{
\dontrun{
list_clinical(Eset)
}
}
