% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Multivariate.R
\name{Multivariate}
\alias{Multivariate}
\title{TCGA Multivariate Analysis}
\usage{
Multivariate(x, Eset, value = 0, factor_list = F, mutation,
  average = "mean", optimal = FALSE, survival = "overall",
  return_df = FALSE, plot_cutpoint = F, coef = F, combined = F,
  stepwise = F, ...)
}
\arguments{
\item{x}{A character vector with 1 or more gene names and/or columns of the clinical patient data}

\item{Eset}{An Expression Set}

\item{value}{Defines the value to subdivide gene expression groups.
Numeric: Devided into two groups (for more genes use value = c(cutpoint gene 1, cutpoint gene 2, ...))
"q": 25 \% quantile, 25-75 \% quantile and 75 \% quantile}

\item{factor_list}{Returns the levels of factors in the analysis}

\item{mutation}{Vector of mutations, that were added using the add_mutation() function}

\item{average}{for more than one gene, how the value of the averaged z-score is calculated, either median or mean}

\item{optimal}{calculate the optimal cutpoint, will overide the value value when numeric, does not work when value = "q"}

\item{survival}{For overall survival = "overall", for disease free survival: survival = "DFS"}

\item{return_df}{Returns the Dataframe that is used for calculatinf the Cox Regression}

\item{plot_cutpoint}{Plot the graph how the optimal cutpoint was calculated, normal survival plot will be REPLACED by surv_cutpoint: Determine the optimal cutpoint for each variable using ’maxstat’}

\item{coef}{Return the coefficients of the multivariate analysis instead of the multivariate object}

\item{combined}{includes a comination of the different genes into the analysis. still in development}

\item{stepwise}{Do a stepwise chosing a model by AIC to get rid of variables that don't play a role in the linear model}

\item{...}{additional variables that can be added}
}
\value{
The summary of the multivariate analysis
}
\description{
TCGA Multivariate Analysis
}
\examples{
\dontrun{
Multivariate(x = c("FOXA2", "GENDER"), Eset = Eset,
average = "median", stepwise=TRUE, coef=TRUE,
optimal=T, plot_cutpoint=F)
}

}
