% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/smoothCoxph_man.R
\name{smoothCoxph_man}
\alias{smoothCoxph_man}
\title{Create a correlation of gene expression to log hazard ratio}
\usage{
smoothCoxph_man(x, Eset, exclude_values, xlimit, ylimit, xlabel, ylabel,
  logrisk = TRUE, survival = "overall", average = "median", ...)
}
\arguments{
\item{x}{1 or multiple Genes, as a character vector list, to analyze}

\item{Eset}{The Expression Set, containing the expression data in columns and the survival indicators in the columns X_OS and X_OS_IND}

\item{exclude_values}{add a vector with following style c("column of clinical data to look for exclusion", "exclude group 1 from this column", "exclude group 2 from this column", ...)}

\item{xlimit}{Manually define the x-axis limit}

\item{ylimit}{Manually define the y-axis limit}

\item{xlabel}{Manually define the x-axis label}

\item{ylabel}{manually define the y-axis label}

\item{logrisk}{don't exactly know... need to check}

\item{survival}{"overall", "DFS"}

\item{average}{values: `mean` or `median`, defines how >1 genes are averaged}

\item{...}{additional variables that can be added}
}
\value{
Plots the log Hazard ratio in correlation to its relative gene expression level using ggplot2
}
\description{
Create a correlation of gene expression to log hazard ratio
}
\examples{
\dontrun{
smoothCoxph_man(x = c("BAMBI", "CD44"), Eset=Eset,
exclude_values=c("pathologic_stage", "Stage II", "Stage I", "Stage IV"),
logrisk=T, average="median")
}

}
