% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TCGADataFelix-package.r
\docType{package}
\name{TCGADataFelix}
\alias{TCGADataFelix}
\alias{TCGADataFelix-package}
\title{TCGADataFelix}
\description{
The package was written to simplify the analysis of the TCGA / GDAC Datasets, as most of the availible packages don't work so far. The analysis is based on the explanation of TriS on Biostars.org: \url{https://www.biostars.org/p/153013/}
}
\details{
The Vignette of this package is a documentation that deals with the first case
}
