% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build_TCGA_Eset.R
\name{build_TCGA_Eset}
\alias{build_TCGA_Eset}
\title{Create an Expression Set from TCGA Data, downloaded from https://gdac.broadinstitute.org/}
\usage{
build_TCGA_Eset(clinical_file, expression_file)
}
\arguments{
\item{clinical_file}{PATH of the clinical data from TCGA, e.g.: "../../../Projects/TCGA-Elisa/Clinical/LAML.clin.merged.txt"}

\item{expression_file}{PATH of the expression data from TCGA, e.g.: "../../../Projects/TCGA-Elisa/RNA/LAML.rnaseqv2__illuminahiseq_rnaseqv2__unc_edu__Level_3__RSEM_genes_normalized__data.data.txt"}
}
\value{
ExpressionSet
}
\description{
Create an Expression Set from TCGA Data, downloaded from https://gdac.broadinstitute.org/
}
\examples{
\dontrun{
build_TCGA_Eset(clinical_file = "../../../Projects/TCGA-Elisa/Clinical/LAML.clin.merged.txt",
expression_file = "../../../Projects/TCGA-Elisa/RNA/LAML.rnaseqv2__illuminahiseq_rnaseqv2__unc_edu__Level_3__RSEM_genes_normalized__data.data.txt")
}

}
