% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fetch_TCGA_mutations.R
\name{fetch_TCGA_mutations}
\alias{fetch_TCGA_mutations}
\title{add a mutation column to the patient data}
\usage{
fetch_TCGA_mutations(path, Eset)
}
\arguments{
\item{path}{Path to the mutation Files downloaded from Firehose TCGA}

\item{Eset}{Eset that will be used}
}
\value{
Returns the Expression Set with added mutation data
}
\description{
add a mutation column to the patient data
}
\examples{
fetch_TCGA_mutations(path = "../../../Projects/TCGA-Elisa/LAML.Mutation_Packager_Calls/", Eset = Eset)



}
