% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Multivariate.R
\name{Multivariate}
\alias{Multivariate}
\title{TCGA Multivariate Analysis}
\usage{
Multivariate(x, Eset, value = 0, factor, factor_list = F, mutation, exclude,
  average = "mean", optimal = FALSE, survival = "overall",
  return_df = FALSE, plot_cutpoint = F, coef = F, ...)
}
\arguments{
\item{x}{A character vector with 1 or more gene names or a column of the clinical patient data}

\item{Eset}{An Expression Set}

\item{value}{Defines the value to subdivide the gene expression groups.
Numeric: Devided into two groups
"q": 25 \% quantile, 25-75 \% quantile and 75 \% quantile}

\item{factor}{Additional covariate for the multivariate analysis}

\item{factor_list}{Returns the levels of factors in the analysis}

\item{mutation}{Vector of mutations, that were added using the add_mutation() function}

\item{exclude}{Factors to exclude, when no gene is entered but another factor from the clinical patient data, it can also be excluded, the order of the to be exlcuded variables does not matter.}

\item{average}{for more than one gene, how the value of the averaged z-score is calculated, either median or mean}

\item{optimal}{calculate the optimal cutpoint, will overide the value value when numeric, does not work when value = "q"}

\item{survival}{For overall survival = "overall", for disease free survival: survival = "DFS"}

\item{return_df}{Returns the Dataframe that is used for calculatinf the Cox Regression}

\item{plot_cutpoint}{Plot the graph how the optimal cutpoint was calculated, normal survival plot will be REPLACED by surv_cutpoint: Determine the optimal cutpoint for each variable using ’maxstat’}

\item{coef}{Return the coefficients of the multivariate analysis instead of the multivariate object}

\item{...}{additional variables that can be added}
}
\value{
The summary of the multivariate analysis
}
\description{
TCGA Multivariate Analysis
}
\examples{
\dontrun{
Multivariate(x = c("FOXA2"), Eset = Eset,
value = 0, factor = "ETHNICITY",
exclude = c("LATINO"), average = "median",
optimal=T, plot_cutpoint=F)
}

}
