% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/survival_adaptable.R
\name{Survival_adaptable}
\alias{Survival_adaptable}
\title{Create a Kaplan Meier Estimator by using a TCGA Dataset}
\usage{
Survival_adaptable(x, Eset, value = 0, additional, exclude, p.val = FALSE,
  xlabel, ylabel, logrisk = TRUE, legend_position, durchschnitt = "mean",
  optimal = FALSE, plot_cutpoint = FALSE, ...)
}
\arguments{
\item{x}{a character vector with 1 or more gene names}

\item{Eset}{An Expression Set}

\item{value}{Defines the value to subdivide the gene expression groups.
Numeric: Devided into two groups
"q": 25 \% quantile, 25-75 \% quantile and 75 \% quantile}

\item{additional}{Additional covariate to subset groups}

\item{exclude}{Factors of the covariates to exclude}

\item{p.val}{display the p-Value on the graph}

\item{xlabel}{User defined x-axis label}

\item{ylabel}{User defined y-axis label}

\item{logrisk}{logrisk?!}

\item{legend_position}{where should the legend be? doesn't work with theme_bw}

\item{durchschnitt}{for more than one gene, how the value of the averaged z-score is calculated, either median or mean}

\item{optimal}{calculate the optimal cutpoint, will overide the value value when numeric, does not work when value = "q"}

\item{plot_cutpoint}{Plot the graph how the optimal cutpoint was calculated, normal survival plot will be REPLACED by surv_cutpoint: Determine the optimal cutpoint for each variable using ’maxstat’}

\item{...}{additional variables that can be added}
}
\value{
A survival Estimator
}
\description{
Create a Kaplan Meier Estimator by using a TCGA Dataset
}
\examples{
\dontrun{
Survival_adaptable(x = c("FOXA2"), Eset = Eset,
value = 0, additional = "pathologic_stage",
exclude = c("Stage IV"), p.val=TRUE,
xlabel="Days", legend_position = "top", durchschnitt = "median",
optimal=T, plot_cutpoint=F)
}
}
